fix: connection

+++ myshoop/back_end/prisma/seed.ts
@@ -28,22 +28,26 @@ type picture = {
 
 async function main() {
 
-  await prisma.publi.deleteMany({})
-  await prisma.urlImage.deleteMany({})
-  await prisma.cart.deleteMany({})
-  await prisma.salesAffiliated.deleteMany({})
-  await prisma.affiliated.deleteMany({})
-  await prisma.payMent.deleteMany({})
-  await prisma.product.deleteMany({})
-  await prisma.categori.deleteMany({})
-  await prisma.storeUser.deleteMany({})
-  await prisma.store.deleteMany({})
-  await prisma.session.deleteMany({})
-  await prisma.addres.deleteMany({})
-  await prisma.user.deleteMany({})
+  const store =   await prisma.store.deleteMany({});
 
-  const data =  await axios.get("https://api.mercadolibre.com/sites/MLB/search?category=MLB3937&listing_type_id=gold_special&condition=new", {headers: {"Authorization":`Bearer ${process.env.TOKEN}` }}) as unknown
-  const list = data as data
+  if(store) return;
+
+  await prisma.publi.deleteMany({});
+  await prisma.urlImage.deleteMany({});
+  await prisma.cart.deleteMany({});
+  await prisma.salesAffiliated.deleteMany({});
+  await prisma.affiliated.deleteMany({});
+  await prisma.payMent.deleteMany({});
+  await prisma.product.deleteMany({});
+  await prisma.categori.deleteMany({});
+  await prisma.storeUser.deleteMany({});
+  await prisma.store.deleteMany({});
+  await prisma.session.deleteMany({});
+  await prisma.addres.deleteMany({});
+  await prisma.user.deleteMany({});
+
+  const data =  await axios.get("https://api.mercadolibre.com/sites/MLB/search?category=MLB3937&listing_type_id=gold_special&condition=new", {headers: {"Authorization":`Bearer ${process.env.TOKEN}` }}) as unknown;
+  const list = data as data;
   const listOu = list.data;
   
   const hashedPassword = await bcrypt.hash("dudududu", 12);
@@ -56,7 +60,7 @@ async function main() {
       urlImage: "https://lh3.googleusercontent.com/ogw/AAEL6sg1I7IEWU2qlQHfNZcz0SiJ4oSa6lZcrCBCta7EpQ=s32-c-mo",
       owner: true
     }
-  })
+  });
 
   const addres = await prisma.addres.create({
     data: {
@@ -67,14 +71,14 @@ async function main() {
       postOfficeCode: 99999999,
       UserId: owner.id
     }
-  })
+  });
 
   const shop = await prisma.store.create({
     data: {
       nameStore: "MYShoop",
       AddresId: addres.id
     }
-  })
+  });
 
   await prisma.storeUser.create({
     data: {
